pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "riddler"
version = "0.1.0"
description = "Generate riddle videos for TikTok"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.59.3",
    "elevenlabs",
    "python-dotenv",
    "opencv-python",
    "numpy",
    "requests",
    "pillow",
    "moviepy",
    "pydantic"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "isort",
    "mypy"
]

[project.scripts]
riddler = "riddler.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/riddler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3 